--- lib/interactify/contract_failure.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module Interactify
-  class ContractFailure < ::Interactor::Failure
-  end
-end
