+++ lib/interactify/contracts/failure.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Interactify
+  module Contracts
+    class Failure < ::Interactor::Failure
+    end
+  end
+end
