+++ spec/lib/interactify/async/job_maker_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe Interactify::JobMaker do
+RSpec.describe Interactify::Async::JobMaker do
   let(:container_klass) { double("ContainerKlass", expected_keys:, promised_keys:, optional_attrs:) }
   let(:optional_attrs) { [] }
   let(:expected_keys) { [] }
