+++ spec/lib/interactify/contracts/helpers.promise_spec.rb
